requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.6
numpy>=1.26
scipy>=1.12
python-rtmidi>=1.5
websockets>=12.0
uvloop>=0.19
pyyaml>=6.0
